[Divya/Chirag] Fixing downfall in replication
After every 20 minutes, the replication used to slow down, and
thereafter, stop completely. This is because the broadcast queue size
increased and it took a lot of time for invalidation of useless messages
to replicate, followed by the goroutine being killed which was
responsible for replication. This commit fixes the concurrency implementation.